Removed nuxt axios and replaced with axios. Created service folder and utility functions instead of methods in vue. reorganised computed properties in dropdown.vue. removed middleware and plugins directory. reorganised layout of vuex store.
