+++ components/fixtures/calendar.vue
@@ -18,6 +18,8 @@
 import toolbar from '~/components/fixtures/filters/toolbar.vue'
 import dropdown from '~/components/fixtures/filters/dropdown.vue'
 import list from '~/components/fixtures/list/list.vue'
+import jsonRetriever from '~/services/jsonRetriever.js'
+import compititonUtility from '~/services/compititonUtility.js'
 import moment from 'moment'
 
 export default {
@@ -42,38 +44,20 @@ export default {
   watch: {
     async selectedCountry() {
       // Gathers the competitions from the JSON data.
-      var competitions = await this.buildCompetitionData()
+      var competitions = await jsonRetriever.buildCompetitionData(this.selectedCountry)
 
       // This removes the unused JSON data.
       // This makes the data smaller and helps with performance.
-      // TODO: This is athletics specific, needs to be moved up to page level.
-      var cleanCompetitions = []
-      competitions.forEach((competition) => {
-        var cleanComp = {
-          id: competition.id,
-          is_demo: competition.is_demo,
-          date: competition.date,
-          full_name: competition.full_name,
-          num_competitors: competition.num_competitors,
-          finish_date: competition.finish_date,
-          type: competition.type,
-          age_groups: competition.age_groups,
-          address: competition.address,
-          entry_link: competition.entry_link,
-          contact_details: competition.contact_details,
-          latitude: competition.latitude,
-          longitude: competition.longitude,
-        }
-        cleanCompetitions.push(cleanComp)
-      })
+      const parentName = this.$parent._name
+      var cleanCompetitions = compititonUtility.cleaner(competitions, parentName.substring(1, parentName.length - 1))
 
       // Sets up all filter items.
       // This needs to be done here as different filter options will be created based
       // on competition data available. One country may have different filter options to another.
-      this.setUpFilters(cleanCompetitions)
+      this.buildMonths(cleanCompetitions)
 
       // Creates a competition tree.
-      this.competitionTree = this.compTreeCreator(cleanCompetitions)
+      this.competitionTree = compititonUtility.treeBuilder(cleanCompetitions, this.months)
     },
   },
   computed: {
@@ -101,99 +85,15 @@ export default {
   created() {
     // Sets the current country to the first country in the list.
     this.$store.dispatch('changeSelectedCountry', this.countries[0])
+
+    // Sorts the counties in alphabetical order.
+    this.countries.sort()
   },
   methods: {
     /**
-     * This is an asynchronous function that collects all json data for a given country.
-     * Then builds a competition tree and sets up the filters on the vue page. This function
-     * should only be called when changing country to display or when creating the vue page.
+     * Creates a list of months based on competition dates.
      */
-    async buildCompetitionData() {
-      var competitions = []
-      // Creates a outer promise (an asynchronous function) that we will
-      // wait to finish.
-      let outerPromise = new Promise((resolve, reject) => {
-        // Creates a lists of inner promises (asynchronous axios calls) based on
-        // the number of JSON data URLs in the selected country.
-        var innerPromises = []
-        this.selectedCountry.urls.forEach((jsonUrl) => {
-          innerPromises.push(this.$axios.get(jsonUrl))
-        })
-
-        // Runs each inner promise, then once all inner promises are complete,
-        // send a message to let the code waiting know the outer promise is done.
-        Promise.all(innerPromises).then(function (results) {
-          results.forEach(function (response) {
-            // Gets the json data and adds it to the list of competitions.
-            competitions.push(response.data)
-          })
-          resolve('done')
-        })
-      })
-
-      // The code waits here for the outer promise to finish before continueing.
-      await outerPromise
-
-      // Flatterns the competitions as after the promise it
-      // contains an array as an element for each json url called.
-      return competitions.flat()
-    },
-    /**
-     * Takes a list of competitions and breaks them down into a tree based on
-     * the available months.
-     */
-    compTreeCreator(competitions) {
-      // Creates month containers and supplies is objects that contain each
-      // available month.
-      var compBranches = []
-      this.months.forEach((month) => {
-        compBranches.push({
-          month: month,
-          competitions: [],
-        })
-      })
-
-      // Loops through each competition
-      competitions.forEach((competition) => {
-        // Do not add the competition if it is a demo fixture and not an actual fixture.
-        if (competition.is_demo) return
-
-        // Loops through each month
-        this.months.forEach((month) => {
-          // If the competition month is the same as the current filtered month
-          // then add it to its branch.
-          if (moment(competition.date).isSame(moment(month, 'MMMM YYYY'), 'month')) {
-            // Finds the branch that matches and adds the competition to its list.
-            var branch = compBranches.find((branch) => branch.month == month)
-            branch.competitions.push(competition)
-          }
-        })
-      })
-
-      // Code section below removes all the branches without any competitions from
-      // the list of compbranches. Its complicated to remove specific elements from
-      // an array in javascript.
-      var tempCompBranches = [...compBranches]
-      var branchesToRemove = []
-      // Loops through the temp array and finds all indexs to remove.
-      var i
-      for (i = 0; i < tempCompBranches.length; i++) {
-        if (tempCompBranches[i].competitions.length == 0) {
-          branchesToRemove.push(i)
-        }
-      }
-      // Loops through temp array in reverse order and removes empty elements.
-      for (var i = branchesToRemove.length - 1; i >= 0; i--) {
-        tempCompBranches.splice(branchesToRemove[i], 1)
-      }
-
-      return tempCompBranches
-    },
-    /**
-     * Sets up the filters to include all available months and countries based off the
-     * provided list of competitions (Untreed).
-     */
-    setUpFilters(competitions) {
+    buildMonths(competitions) {
       // Adds the current month to list of available months.
       this.$store.dispatch('changeSelectedMonth', moment().format('MMMM YYYY'))
       if (!this.months.includes(this.selectedMonth)) {
@@ -208,9 +108,6 @@ export default {
           this.months.push(month)
         }
       })
-
-      // Sorts the counties in alphabetical order.
-      this.countries.sort()
 
       // Sorts the months in calendar order.
       this.months.sort((x, y) => moment(x, 'MMMM YYYY') - moment(y, 'MMMM YYYY'))
