+++ package.json
@@ -11,8 +11,8 @@
   },
   "dependencies": {
     "@nuxt/typescript-runtime": "^2.0.0",
-    "@nuxtjs/axios": "^5.12.2",
     "@nuxtjs/pwa": "^3.0.2",
+    "axios": "^0.20.0",
     "core-js": "^3.6.5",
     "moment": "^2.29.0",
     "nuxt": "^2.14.5"
